api-gateway/requirements.txt
Django>=4.0
requests
djangorestframework
